pyproject.toml
[project]
name = "bq-spark-bm"
version = "0.1.0"
description = "BigQuery Spark benchmark"
readme = "README.md"
authors = [
    { name = "bemihai", email = "mihaita.berbec@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyspark>=3.5.5",
    "python-dotenv==1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

 [tool.hatch.build.targets.wheel]
 packages = ["src"]
